+++ packages/web-framework/server/auth-manager.js
@@ -24,8 +24,6 @@ const authMethodEnum = tools.enum( {
     OPENID_GOOGLE: [ "openid-google", "openid-google", "Authentication to Google using OpenID Connect." ]
 } );
 
-module.exports.authMethod = authMethodEnum;
-
 /**
  * @class AuthManager
  * @public
@@ -35,13 +33,14 @@ class AuthManager {
     /** @type {SettingsAuth} */
     #authSettings = {
         enabledMethods: [],
+        local: {
+            username: undefined,
+            password: undefined,
+            enabled: false
+        },
         oauth2: {}
     };
-    #localAuthentication = {
-        username: undefined,
-        password: undefined,
-        enabled: false
-    };
+    #googleOauth2Configuration = {};
 
     /**
      * @constructor
@@ -54,9 +53,10 @@ class AuthManager {
 
         if ( this.#authSettings.enabledMethods.includes( authMethodEnum.LOCAL ) ) {
             // TODO: For testing purposes only! Implement real local auth later!
-            this.#localAuthentication.username = "admin";
-            this.#localAuthentication.password = "admin";
-            this.#localAuthentication.enabled = true;
+            this.#authSettings.local = this.#authSettings.local || {};
+            this.#authSettings.local.username = "admin";
+            this.#authSettings.local.password = "admin";
+            this.#authSettings.local.enabled = true;
         }
 
         if ( this.#authSettings.enabledMethods.includes( authMethodEnum.OPENID_GOOGLE ) ) {
@@ -82,7 +82,7 @@ class AuthManager {
             if ( this.#authSettings.enabledMethods.includes( authMethodEnum.OPENID_GOOGLE ) ) {
                 promises.push( this.#initializeOpenIDGoogle() );
             }
-            return Promise.all( promises ).then( ( result ) => {
+            return Promise.all( promises ).then( () => {
                 resolve();
             } ).catch( ( error ) => {
                 reject( exceptions.raise( error ) );
@@ -91,27 +91,89 @@ class AuthManager {
     }
 
     /**
-     * Used to verify the local authentication of a request.
+     * Used to check whether the specified authentication method is enabled.
      *
      * @method
-     * @param {string} username
-     * @param {string} password
+     * @param {TiAuthMethod} authMethod
      * @returns {boolean}
      * @public
      */
-    localAuthentication( username, password ) {
-        // TODO: Implement this!
-        if ( this.#localAuthentication.enabled === true ) {
-            return ( username === this.#localAuthentication.username && password === this.#localAuthentication.password );
+    isEnabled( authMethod ) {
+        return this.#authSettings.enabledMethods.includes( authMethod );
+    }
+
+    /**
+     * Used to authenticate a user via the specified authentication method.
+     *
+     * @method
+     * @param {TiAuthMethod} authMethod
+     * @param {Object} authDetails
+     * @returns {Promise}
+     * @throws {Exception} If the authentication method is not recognized or enabled.
+     * @public
+     */
+    authenticate( authMethod, authDetails ) {
+        switch ( authMethod ) {
+            case authMethodEnum.LOCAL:
+                return this.#localAuthentication( authDetails.username, authDetails.password );
+            case authMethodEnum.OPENID_GOOGLE:
+                return this.#openIDGoogleAuthentication( authDetails.baseUrl );
+            default: {
+                let exception = exceptions.raise( exceptions.exceptionCode.E_SEC_UNRECOGNIZED_AUTH_METHOD );
+                exception.httpCode = exceptions.httpCode.C_401;
+                throw exception;
+            }
+        }
+    }
+
+    /**
+     * Used to get the callback URL for Google authentication.
+     * <br/>
+     * NOTE: This method is only available if the Google authentication method is enabled.
+     *
+     * @method
+     * @returns {string}
+     * @throws {Exception} If the Google authentication method is not enabled.
+     * @public
+     */
+    getGoogleCallbackUrl() {
+        if ( this.#authSettings.enabledMethods.includes( authMethodEnum.OPENID_GOOGLE ) ) {
+            return this.#authSettings.oauth2.google.callbackUrl;
         } else {
-            return false;
+            let exception = exceptions.raise( exceptions.exceptionCode.E_SEC_UNRECOGNIZED_AUTH_METHOD );
+            exception.httpCode = exceptions.httpCode.C_401;
+            throw exception;
         }
     }
 
     /* Private interface */
 
     /**
+     * Used to verify the local authentication of a request.
+     *
+     * @method
+     * @param {string} username
+     * @param {string} password
+     * @returns {Promise<boolean>}
+     * @private
+     */
+    #localAuthentication( username, password ) {
+        return new Promise( ( resolve, reject ) => {
+            // TODO: Implement this!
+            if ( this.#authSettings.local.enabled === true ) {
+                resolve( username === this.#authSettings.local.username && password === this.#authSettings.local.password );
+            } else {
+                let exception = exceptions.raise( exceptions.exceptionCode.E_SEC_UNAUTHORIZED_ACCESS );
+                exception.httpCode = exceptions.httpCode.C_401;
+                reject( exception );
+            }
+        } );
+    }
+
+    /**
      * Used to initialize the OpenID Connect client for Google authentication.
+     * <br/>
+     * NOTE: Guide available here https://developers.google.com/identity/openid-connect/openid-connect
      *
      * @method
      * @returns {Promise}
@@ -120,7 +182,10 @@ class AuthManager {
     #initializeOpenIDGoogle() {
         return new Promise( ( resolve, reject ) => {
             // TODO: Implement this!
-            openidClient.discovery( new URL( "" ), this.#authSettings.oauth2.clientID, this.#authSettings.oauth2.clientSecret ).then( () => {
+            const googleDiscoveryUrl = "https://accounts.google.com/.well-known/openid-configuration";
+            const googleSettings = this.#authSettings.oauth2.google;
+            openidClient.discovery( new URL( googleDiscoveryUrl ), googleSettings.clientID, googleSettings.clientSecret ).then( ( configuration ) => {
+                this.#googleOauth2Configuration = configuration;
                 resolve();
             } ).catch( ( error ) => {
                 reject( error );
@@ -128,6 +193,38 @@ class AuthManager {
         } );
     }
 
+    /**
+     * Used to perform the actual OpenID Connect authentication.
+     *
+     * @method
+     * @param {string} baseUrl
+     * @returns {Promise<URL>}
+     * @private
+     */
+    #openIDGoogleAuthentication( baseUrl ) {
+        return new Promise( ( resolve, reject ) => {
+            const codeVerifier = openidClient.randomPKCECodeVerifier();
+            openidClient.calculatePKCECodeChallenge( codeVerifier ).then( ( codeChallenge ) => {
+                const parameters = {
+                    redirect_uri: `${ baseUrl }${ this.#authSettings.oauth2.google.callbackUrl }`,
+                    response_type: "code",
+                    client_id: this.#authSettings.oauth2.google.clientID,
+                    scope: "openid email profile",
+                    code_challenge: codeChallenge,
+                    code_challenge_method: "S256"
+                };
+                if ( !this.#googleOauth2Configuration.serverMetadata().supportsPKCE() ) {
+                    parameters.state = openidClient.randomState();
+                }
+                const redirectTo = openidClient.buildAuthorizationUrl( this.#googleOauth2Configuration, parameters );
+                resolve( redirectTo );
+            } ).catch( ( error ) => {
+                reject( exceptions.raise( error ) );
+            } );
+        } );
+    }
+
 }
 
 module.exports = AuthManager;
+module.exports.authMethod = authMethodEnum;